requirements.txt
aiohttp==3.11.18
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.4.2
azure-cosmos==4.5.1
azure-storage-blob
python-jose[cryptography]==3.3.0
gunicorn==21.2.0
pyyaml
openai
PyMuPDF
python-docx
reportlab
PyGithub
autogen
sqlalchemy